deleted example stuff
